move unification and tarjan algorithm to algorithms subdirectory
